fix(release): clarify asset filename normalization
