+++ pkg/cmd/release/shared/upload_test.go
@@ -9,6 +9,7 @@ import (
 	"testing"
 
 	"github.com/cli/cli/v2/internal/safeurl"
+	"github.com/stretchr/testify/assert"
 )
 
 func Test_typeForFilename(t *testing.T) {
@@ -73,6 +74,51 @@ func Test_typeForFilename(t *testing.T) {
 			if got := typeForFilename(tt.file); got != tt.want {
 				t.Errorf("typeForFilename() = %v, want %v", got, tt.want)
 			}
+		})
+	}
+}
+
+func Test_SanitizeFileName(t *testing.T) {
+	tests := []struct {
+		name     string
+		expected string
+	}{
+		{
+			name:     "foo",
+			expected: "foo",
+		},
+		{
+			name:     "foo bar",
+			expected: "foo.bar",
+		},
+		{
+			name:     ".foo",
+			expected: "default.foo",
+		},
+		{
+			name:     "Foo bar",
+			expected: "Foo.bar",
+		},
+		{
+			name:     "Hello, दुनिया",
+			expected: "default.Hello",
+		},
+		{
+			name:     "this+has+plusses.jpg",
+			expected: "this+has+plusses.jpg",
+		},
+		{
+			name:     "this@has@at@signs.jpg",
+			expected: "this@has@at@signs.jpg",
+		},
+		{
+			name:     "façade.exposé",
+			expected: "facade.expose",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			assert.Equal(t, tt.expected, SanitizeFileName(tt.name))
 		})
 	}
 }
