+++ pkg/cmd/release/upload/upload_test.go
@@ -7,57 +7,11 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/cli/cli/v2/pkg/cmd/release/shared"
 	"github.com/cli/cli/v2/pkg/cmdutil"
 	"github.com/cli/cli/v2/pkg/iostreams"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-func Test_SanitizeFileName(t *testing.T) {
-	tests := []struct {
-		name     string
-		expected string
-	}{
-		{
-			name:     "foo",
-			expected: "foo",
-		},
-		{
-			name:     "foo bar",
-			expected: "foo.bar",
-		},
-		{
-			name:     ".foo",
-			expected: "default.foo",
-		},
-		{
-			name:     "Foo bar",
-			expected: "Foo.bar",
-		},
-		{
-			name:     "Hello, दुनिया",
-			expected: "default.Hello",
-		},
-		{
-			name:     "this+has+plusses.jpg",
-			expected: "this+has+plusses.jpg",
-		},
-		{
-			name:     "this@has@at@signs.jpg",
-			expected: "this@has@at@signs.jpg",
-		},
-		{
-			name:     "façade.exposé",
-			expected: "facade.expose",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.expected, shared.SanitizeFileName(tt.name))
-		})
-	}
-}
 
 func Test_NewCmdUpload_rejectsDuplicateAssetFilenames(t *testing.T) {
 	firstDir := t.TempDir()
